requirements.txt
sounddevice
numpy
scipy
customtkinter
faster-whisper
cryptography
matplotlib
fpdf2
plyer
